+++ src/sub-clients/events/client/EventsClient.test.pact.ts
@@ -1,19 +1,11 @@
 import { ApiBoclipsClient } from '../../../ApiBoclipsClient';
 import { provider } from '../../../pact-support/pactSetup';
 import { withClients } from '../../../pact-support/pactTestWrapper';
-import {
-  CollectionFactory,
-  FakeBoclipsClient,
-  isATestClient,
-} from '../../../test-support';
+import { CollectionFactory, FakeBoclipsClient, isATestClient } from '../../../test-support';
 import { Link } from '../../common/model/LinkEntity';
-import { CollectionInteractedWithRequest } from '../model/CollectionInteractedWithRequest';
+import { CollectionInteractedWithRequest, CollectionInteractionType } from '../model/CollectionInteractedWithRequest';
 import { PageRenderedRequest } from '../model/PageRenderedRequest';
-import {
-  collectionID,
-  trackCollectionInteraction,
-  trackPageRendered,
-} from '../pact/EventsInteractions';
+import { collectionID, trackCollectionInteraction, trackPageRendered } from '../pact/EventsInteractions';
 
 describe('EventsClient', () => {
   withClients(
@@ -55,7 +47,7 @@ describe('EventsClient', () => {
 
       it(`can track collection interaction events`, async () => {
         const request: CollectionInteractedWithRequest = {
-          subtype: 'NAVIGATE_TO_COLLECTION_DETAILS',
+          subtype: CollectionInteractionType.NAVIGATE_TO_COLLECTION_DETAILS,
         };
 
         const collection = CollectionFactory.sample({