+++ src/sub-clients/events/model/CollectionInteractedWithRequest.ts
@@ -1,5 +1,9 @@
 import { EventRequest } from './EventRequest';
 
 export interface CollectionInteractedWithRequest extends EventRequest {
-  subtype: string;
+  subtype: CollectionInteractionType;
+}
+
+export enum CollectionInteractionType {
+  NAVIGATE_TO_COLLECTION_DETAILS = 'NAVIGATE_TO_COLLECTION_DETAILS',
 }